requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.7
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==19.2
psycopg2==2.8.4
pycparser==2.19
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
mysqlclient
